# Conflicts:
#	src/CMakeLists.txt
#	src/genomeRefactored/Genome.cpp
#	src/genomeRefactored/GenomeIndex_find.cpp
#	src/genomeRefactored/SJDB.cpp
#	src/genomeRefactored/SuffixArray.cpp
#	src/genomeRefactored/SuffixArrayKMerMap.h
#	src/readAlignRefactored/Stitching.cpp
#	src/readAlignRefactored/Transcript.cpp
#	src/utils/PackedIndexArray.h
#	src/utilsRefactored/seqFunctions.cpp

+++ src/genomeRefactored/Genome.cpp
@@ -81,7 +81,7 @@ namespace RefactorProcessing {
 
         if (!currentChr.empty()) {
 
-            size_t currentPos = static_cast<int64_t>(currentSeq.tellp());
+            size_t currentPos = static_cast<size_t>(currentSeq.tellp());
             size_t chrLen = currentPos - currentStart;
             size_t paddingLen = ((chrLen + (1 << binSizeLog_) - 1) & ~((1 << binSizeLog_) - 1)) - chrLen;
             for (size_t i = 0; i < paddingLen; ++i) {
